feat: Pass stream processor in from client lib
